+++ Samples~/Scripts/SaintsEditor/Testing/LayoutFoldout.cs
@@ -4,17 +4,17 @@ namespace SaintsField.Samples.Scripts.SaintsEditor.Testing
 {
     public class LayoutFoldout : SaintsMonoBehaviour
     {
-        [LayoutStart("Grouped", ELayout.Foldout)]
+        [LayoutStart("Foldout", ELayout.Foldout)]
         public string t1;
         public string t2;
         public string t3;
 
-        [LayoutStart("GroupedBox", ELayout.Foldout |  ELayout.TitleBox)]
+        [LayoutStart("Foldout <color=DeepSkyBlue>Box", ELayout.Foldout |  ELayout.TitleBox)]
         public string b1;
         public string b2;
         public string b3;
 
-        [LayoutStart("GroupedOut", ELayout.Foldout | ELayout.TitleOut)]
+        [LayoutStart("<icon=LensFlare Gizmo/>Foldout", ELayout.Foldout | ELayout.TitleOut)]
         public string o1;
         public string o2;
         public string o3;